requirements.txt
gunicorn==19.6.0
Flask==0.11.1
WTForms==2.1
tox==2.5.0
flake8==3.2.0
pyflakes==1.3.0
Flask-WTF==0.13.1
pytidylib==0.3.2
